PR Feedback: move the DebuggerRequestResult to controls, and add copyright header
